Add files via upload
Real meteor reflection log files from Ellensburg, WA. The file name is the start date and time, Pacific time.  A . means five seconds with no signal. M means the start of a meteor and N means the end, and a character represents the amplitude on a scale of 0 to F (with 3 being the threshold), and a character represents 10ms of signal. Anything under about 80ms is spurious noise. F means the signal is clipped.